pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "soccer-market-maker"
version = "0.1.0"
description = "Soccer market making bot for Kalshi and Polymarket"
authors = [{name = "Soccer MM Team"}]
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "pydantic>=2,<3",
    "httpx>=0.27,<1",
    "websockets>=12,<13",
    "cryptography>=41,<45",
    "orjson>=3.9,<4",
    "pyyaml>=6.0,<7",
]

[project.optional-dependencies]
dev = [
    "ruff>=0.4",
    "mypy>=1.8",
    "pytest>=8",
    "pytest-asyncio>=0.23",
    "pre-commit>=3.7",
]

[tool.ruff]
line-length = 100
target-version = "py312"

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W", "UP", "B", "C4", "PIE", "T20"]
ignore = ["E501", "N818"]

[tool.mypy]
python_version = "3.12"
strict = true
warn_return_any = true
warn_unused_configs = true